remove public access from implementation details
