+++ src/lib.rs
@@ -1,14 +1,14 @@
 #[macro_use]
 extern crate diesel;
 
-pub mod ewkb;
-pub mod linestring;
-pub mod multiline;
-pub mod multipoint;
-pub mod multipolygon;
+mod ewkb;
+mod geometrycollection;
+mod linestring;
+mod multiline;
+mod multipoint;
+mod multipolygon;
 pub mod operators;
-pub mod points;
-pub mod polygon;
+mod points;
+mod polygon;
 pub mod sql_types;
 pub mod types;
-pub mod geometrycollection;
