+++ src/linestring.rs
@@ -8,18 +8,11 @@ use diesel::{
     serialize::{self, IsNull, Output, ToSql},
 };
 
-use crate::ewkb::{
+use crate::{ewkb::{
     read_ewkb_header, write_ewkb_header, EwkbSerializable, GeometryType, BIG_ENDIAN,
-};
-use crate::points::{read_point_coordinates, write_point_coordinates, Dimension, PointT};
+}, types::{LineString, PointT}};
+use crate::points::{read_point_coordinates, write_point_coordinates, Dimension};
 use crate::sql_types::*;
-
-#[derive(Clone, Debug, PartialEq, FromSqlRow, AsExpression)]
-#[diesel(sql_type = Geometry)]
-pub struct LineString<T> {
-    pub points: Vec<T>,
-    pub srid: Option<u32>,
-}
 
 impl<T> EwkbSerializable for LineString<T>
 where
